changelog till last merge request from Sep 18, 2019:
added initialization of the timestamp, when None was passed and parsing datetime to seconds. (MissingMatchPathValueDetector, EnhancedNewMatchPathValueComboDetector, HistogramAnalysis, TimeCorrelationDetector, TimestampCorrectionFilters ...)
corrected multiple implementation errors, so the classes work as intended. (MissingMatchPathValueDetector, HistogramAnalysis, TimeCorrelationViolationDetector ...)
PathDependentHistogramAnalysis now creates multiple reports.
added the original log line feature. (output of the original log line at an event)
improved the wrong count of lines in the output of the Events.
changed the matchObjects for multiple parsing models to be more useful data.
re-implemented the MultiLocaleDateTimeModelElement.
fixed the functionality of some parsing models and adapted them to Python 3.
added input-validation to multiple classes and handled error inputs without the crashing of the AMiner.

changelog from Sep 18, 2019 till Mar 19, 2020:

fixed memory leaks in the TimeCorrelationViolationDetector
added properties for the configuration of maximal permitted resource usage (RAM and CPU Usage)
completely re-implemented the AMinerRemoteControl. Now it is based on an whitelist of allowed methods.
renewed the man pages for the AMinerRemoteControl and the AMiner.
implemented the JsonConverterHandler and adapted all current Detectors to save the data in a standardized way.
fixed multiple bugs with the conversion of timestamps and datetimes.
implemented the CI/CD .gitlab-ci.yml
added statement and branch coverage tests to the CI/CD
fixed over 1400 Issues in the SonarQube to meet requirements of the PEP8 Standard
fixed errors in the DelimitedDataModelElement