Refactored multiple variable names and implemented a default __init__ method in the AtomHandlerInterface.
The AtomHandlerInterface.__init__() method now contains (unfinished yet) all possible arguments for the analysis components. This ensures,
that duplicate variable names are removed/renamed. Also type checks of the arguements may be done in future in the superclass. The method
sets all non-None attributes to the self object. To ensure that no "defined outside __init__()" issues are found, some attributes still need
to be set to None.