Feature jsonstringparser (#1167)
* Added experimental JsonStringModelElement

This commit adds an experimental JsonStringModelElement which is
supposed to be much faster than the JsonModelElement. This parser
uses the orjson-parser which reduces complexity with moderate speed.

This commit also modifies the build_docker.sh to support different
container-engines e.g. podman

* Support for strict-mode in JsonStringModelElement

This commit adds support for strict-mode to the JsonStringModelElement.
It also adds orjson and vim to the aecid-testsuite-Dockerfile.

* Added initial testfile for JsonStringModelElement

* add documentation for JsonStringModelElement

* Improved logging for JsonStringModelElement

* Handle "null"-values in JsonStringModelElement

For now the JsonStringModelElement will change "null"-values to
empty strings(bytes).

* Add ignore_null param to JsonStringModelElement

This commmits adds the parameter "ignore_null" to
JsonStringModelElement. This parameter allows users
to ignore all json-keys which have the value "null".
If ignore_null is set to false, the subparser will be called with
b'' (empty string) instead of None

* Docs: ignore_null to JsonStringModelElement

This commit adds the parameter "ignore_null" to the documentation of the
JsonStringModelElement.

* Add tests for ignore_null(JsonStringModelElement)

This commit adds testcases for the ignore_null parameter of the
JsonStringModelElement.

* Add test for JsonAccesObject

This commit adds basic tests for the JsonAccessObject.

* updated version to 2.6.0

* Docs: Added example for JsonStringModelElement

This commit adds a code sample to the documentation of the
JsonStringModelElement.

* DeepSource: indentation fixes

* JsonStringModelElement: missed __init__-call

In JsonStringModelElement __init__ of the superclass is not called in
the constructor. This commit adds the __init__-call and fixes the
unit-test.

* DeepSource: fixed whitespaces

* merged if statement in ModelElementInterface

* JsonAccessObject: add debug-mode

* DeepSource: fixed whitespaces

* JsonAccessObject: fixed debug-mode

* JsonAccessObject: skipped FLK-E228

* Removed multiple empty lines

* Shortend some lines in JsonStringModelElement

* Shortend some lines in YamlConfig

* Add docstrings to JsonStringModelElement

* replaced docstring with comment in JsonStringModelElement

* replaced format-strins in JsonStringModelElement

* DeepSource: fixed whitespaces, format-strings and too long lines

* DeepSource: fixed too long line in JsonStringModelElement

* Added initial pre-commit-config

* skipped test for f-string in metadata.py

* Fixed strict types in JsonStringModelElement

* Fixed Problem with Lists in JsonAccessObject

* Add flake8 to pre-commit

* Fixed docstrings for JsonStringModelElement

* fixed format-error in ModelElementInterface

Co-authored-by: Wolfgang Hotwagner <wolfgang.hotwagner@ait.ac.at>